pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "faiss-cpu>=1.10.0",
    "google-generativeai>=0.8.4",
    "numpy>=2.2.4",
    "requests>=2.32.3",
    "streamlit>=1.44.1",
    "trafilatura>=2.0.0",
]
